+++ .env.schema
@@ -1,5 +1,5 @@
 PORT=
-TOKEN_SECRET
+TOKEN_SECRET=
 
 # DB Connection
 DB_NAME=
