+++ package.json
@@ -29,6 +29,7 @@
     "@babel/core": "^7.8.3",
     "@babel/polyfill": "^7.8.3",
     "@babel/preset-env": "^7.8.3",
+    "babel-eslint": "^10.0.3",
     "babel-loader": "^8.0.6",
     "clean-webpack-plugin": "^3.0.0",
     "eslint": "^6.8.0",
@@ -42,6 +43,7 @@
     "husky": "^4.2.1",
     "lint-staged": "^10.0.7",
     "nodemon-webpack-plugin": "^4.2.2",
+    "prettier": "^1.19.1",
     "webpack": "^4.41.5",
     "webpack-cli": "^3.3.10",
     "webpack-dev-server": "^3.10.1",
