+++ src/components/auth/service.js
@@ -1,6 +1,6 @@
 import { sign } from 'jsonwebtoken';
 
-import { TOKEN_SECRET } from "../../config/env";
+import { TOKEN_SECRET } from '../../config/env';
 
 export const createToken = (data, expireTime) =>
   sign(data, TOKEN_SECRET, {
