+++ src/components/users/service.js
@@ -1,11 +1,13 @@
-import { genRandomString, getHmac } from "../../utils/hash";
+import { genRandomString, getHmac } from '../../utils/hash';
 import UserModel from './model';
-import {DB_SALT} from "../../config/env";
+import { DB_SALT } from '../../config/env';
 
-export const isEmailTaken = async email =>
-  (await UserModel.findOne({ email })) !== null;
+export const isEmailTaken = async email => {
+  const user = await UserModel.findOne({ email });
+  return user !== null;
+};
 
-export const createUser = ({email, password, name}) => {
+export const createUser = ({ email, password, name }) => {
   const salt = genRandomString(32);
 
   return UserModel.create({
