+++ src/db/index.js
@@ -1,16 +1,16 @@
 import mongoose from 'mongoose';
 
-import { DB_URL, DB_NAME } from '../config/env'
+import { DB_URL, DB_NAME } from '../config/env';
 
 export default () =>
-    mongoose
-        .connect(DB_URL, {
-            dbName: DB_NAME,
-            useNewUrlParser: true,
-            useCreateIndex: true,
-            useUnifiedTopology: true,
-            retryWrites: true,
-            w: 'majority',
-        })
-        .then(() => console.log('Mongo connected')) // eslint-disable-line no-console
-        .catch(err => console.log(err)); // eslint-disable-line no-console
+  mongoose
+    .connect(DB_URL, {
+      dbName: DB_NAME,
+      useNewUrlParser: true,
+      useCreateIndex: true,
+      useUnifiedTopology: true,
+      retryWrites: true,
+      w: 'majority',
+    })
+    .then(() => console.log('Mongo connected')) // eslint-disable-line no-console
+    .catch(err => console.log(err)); // eslint-disable-line no-console
