+++ src/index.js
@@ -5,10 +5,17 @@ import createDbConnection from './db';
 
 const app = express();
 
-createDbConnection();
-
 app.use(json());
-
 app.use('/api', routes);
 
-app.listen(PORT, () => console.log('Server listening on:', PORT));
+(async () => {
+  try {
+    await createDbConnection();
+
+    // eslint-disable-next-line no-console
+    app.listen(PORT, () => console.log('Server listening on:', PORT));
+  } catch (e) {
+    // eslint-disable-next-line no-console
+    console.log('Something went wrong');
+  }
+})();
